+++ helianto-partner/src/main/java/org/helianto/partner/form/CompositePartnerForm.java
@@ -74,7 +74,7 @@ public class CompositePartnerForm
 		this.parent = parent;
 	}
 	
-	public long getParentId() {
+	public int getParentId() {
 		if (getParent()!=null) {
 			return getParent().getId();
 		}
